requirements.txt
flask
transformers
torch
Pillow
gunicorn
flasgger
